move dataclasses.py to descriptions.py
